requirements.txt
torch
torchvision
torchmetrics
numpy
opencv-python
matplotlib
scipy
scikit-image
wandb
timm
pillow
pycocotools
PyYAML
